+++ src/main/resources/data.sql
@@ -1,5 +1,5 @@
 -- 상위 메뉴 1
-INSERT INTO algorithm (lv, name, try_count)
-VALUES (2, '타겟넘버', 1);
-INSERT INTO algorithm (lv, name, try_count)
-VALUES (3, '최단 맵 거리', 1);
+INSERT INTO solve (title, level, solved_count, tag, description)
+VALUES ('타겟넘버', 2, 1, 'dfs', '어렵다');
+INSERT INTO solve (title, level, solved_count, tag, description)
+VALUES ('최단 맵 거리', 3, 3, 'bfs', '쉽다');
